refactor: implementação do projeto com typescript, orientação a objetos e DDD
